requirements.txt
streamlit>=1.36.0
beautifulsoup4>=4.12.3
requests>=2.32.3
pydantic>=2.7.4
tenacity>=8.3.0
python-dotenv>=1.0.1
scikit-learn>=1.5.1
openai>=1.51.0
tiktoken>=0.7.0
numpy>=1.26.4
pandas>=2.2.2
